Move framework files into an importable module
