+++ a_simple_dungeon/__init__.py
@@ -1,0 +1,1 @@
+from . import classes, functions, worldmap
